+++ async_upnp_client/aiohttp.py
@@ -7,6 +7,7 @@ import socket
 
 import aiohttp
 import aiohttp.web
+import async_timeout
 
 from async_upnp_client import UpnpRequester
 from async_upnp_client import UpnpEventHandler
@@ -31,21 +32,26 @@ def get_local_ip(target_host=None) -> str:
 class AiohttpRequester(UpnpRequester):
     """Standard AioHttpUpnpRequester, to be used with UpnpFactory."""
 
+    def __init__(self, timeout=5):
+        """Initializer."""
+        self._timeout = timeout
+
     async def async_do_http_request(self, method, url, headers=None, body=None, body_type='text'):
         """Do a HTTP request."""
         # pylint: disable=too-many-arguments
 
-        async with aiohttp.ClientSession() as session:
-            async with session.request(method, url, headers=headers, data=body) as response:
-                status = response.status
-                headers = response.headers
+        async with async_timeout.timeout(self._timeout):
+            async with aiohttp.ClientSession() as session:
+                async with session.request(method, url, headers=headers, data=body) as response:
+                    status = response.status
+                    headers = response.headers
 
-                if body_type == 'text':
-                    body = await response.text()
-                elif body_type == 'raw':
-                    body = await response.read()
-                elif body_type == 'ignore':
-                    body = None
+                    if body_type == 'text':
+                        body = await response.text()
+                    elif body_type == 'raw':
+                        body = await response.read()
+                    elif body_type == 'ignore':
+                        body = None
 
         return status, headers, body
 
@@ -57,10 +63,11 @@ class AiohttpSessionRequester(UpnpRequester):
     With pluggable session.
     """
 
-    def __init__(self, session, with_sleep=False):
+    def __init__(self, session, with_sleep=False, timeout=5):
         """Initializer."""
         self._session = session
         self._with_sleep = with_sleep
+        self._timeout = timeout
 
     async def async_do_http_request(self, method, url, headers=None, body=None, body_type='text'):
         """Do a HTTP request."""
@@ -69,16 +76,17 @@ class AiohttpSessionRequester(UpnpRequester):
         if self._with_sleep:
             await asyncio.sleep(0.01)
 
-        async with self._session.request(method, url, headers=headers, data=body) as response:
-            status = response.status
-            headers = response.headers
+        async with async_timeout.timeout(self._timeout):
+            async with self._session.request(method, url, headers=headers, data=body) as response:
+                status = response.status
+                headers = response.headers
 
-            if body_type == 'text':
-                body = await response.text()
-            elif body_type == 'raw':
-                body = await response.read()
-            elif body_type == 'ignore':
-                body = None
+                if body_type == 'text':
+                    body = await response.text()
+                elif body_type == 'raw':
+                    body = await response.read()
+                elif body_type == 'ignore':
+                    body = None
 
         return status, headers, body